Updated preprocessor for Feather m0 I2C address choice
Modified the preprocessor ifdefs for I2C address choice to exclude
feather m0 boards from arduino zero error, and leave the I2C address as
the default for feathers.

As the feather has the same processor as the zero, and seems to trigger
on the ifdef ARDUINO_SAMD_ZERO this directive would trigger for
that board also. However it is the Zero's debug chip that requires the
address change, not the processor. Confirmed this code as working on my
feather m0.